refactor(#97): restructure API endpoints to follow RESTful conventions
