+++ apps/backend/core/src/main/java/com/schemafy/core/erd/controller/ColumnController.java
@@ -1,7 +1,5 @@
 package com.schemafy.core.erd.controller;
 
-import java.util.List;
-
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -48,15 +46,6 @@ public class ColumnController {
     public Mono<BaseResponse<ColumnResponse>> getColumn(
             @PathVariable String columnId) {
         return columnService.getColumn(columnId)
-                .map(BaseResponse::success);
-    }
-
-    @PreAuthorize("hasAnyRole('OWNER','ADMIN','EDITOR','COMMENTER','VIEWER')")
-    @GetMapping("/columns/table/{tableId}")
-    public Mono<BaseResponse<List<ColumnResponse>>> getColumnsByTableId(
-            @PathVariable String tableId) {
-        return columnService.getColumnsByTableId(tableId)
-                .collectList()
                 .map(BaseResponse::success);
     }
