+++ apps/backend/core/src/main/java/com/schemafy/core/erd/controller/MemoController.java
@@ -56,15 +56,6 @@ public class MemoController {
                 .map(BaseResponse::success);
     }
 
-    @PreAuthorize("hasAnyRole('OWNER','ADMIN','EDITOR','COMMENTER','VIEWER')")
-    @GetMapping("/memos/schema/{schemaId}")
-    public Mono<BaseResponse<List<MemoResponse>>> getMemosBySchemaId(
-            @PathVariable String schemaId) {
-        return memoService.getMemosBySchemaId(schemaId)
-                .collectList()
-                .map(BaseResponse::success);
-    }
-
     @PreAuthorize("hasAnyRole('OWNER','ADMIN','EDITOR','COMMENTER')")
     @PutMapping("/memos/{memoId}")
     public Mono<BaseResponse<MemoResponse>> updateMemo(
