+++ apps/backend/core/src/main/java/com/schemafy/core/erd/controller/TableController.java
@@ -19,8 +19,16 @@ import com.schemafy.core.common.exception.ErrorCode;
 import com.schemafy.core.common.type.BaseResponse;
 import com.schemafy.core.erd.controller.dto.request.CreateTableRequestWithExtra;
 import com.schemafy.core.erd.controller.dto.response.AffectedMappingResponse;
+import com.schemafy.core.erd.controller.dto.response.ColumnResponse;
+import com.schemafy.core.erd.controller.dto.response.ConstraintResponse;
+import com.schemafy.core.erd.controller.dto.response.IndexResponse;
+import com.schemafy.core.erd.controller.dto.response.RelationshipResponse;
 import com.schemafy.core.erd.controller.dto.response.TableDetailResponse;
 import com.schemafy.core.erd.controller.dto.response.TableResponse;
+import com.schemafy.core.erd.service.ColumnService;
+import com.schemafy.core.erd.service.ConstraintService;
+import com.schemafy.core.erd.service.IndexService;
+import com.schemafy.core.erd.service.RelationshipService;
 import com.schemafy.core.erd.service.TableService;
 
 import lombok.AllArgsConstructor;
@@ -37,6 +45,10 @@ import validation.Validation.DeleteTableRequest;
 public class TableController {
 
     private final TableService tableService;
+    private final ColumnService columnService;
+    private final RelationshipService relationshipService;
+    private final IndexService indexService;
+    private final ConstraintService constraintService;
 
     @PreAuthorize("hasAnyRole('OWNER','ADMIN','EDITOR')")
     @PostMapping("/tables")
@@ -59,10 +71,37 @@ public class TableController {
     }
 
     @PreAuthorize("hasAnyRole('OWNER','ADMIN','EDITOR','COMMENTER','VIEWER')")
-    @GetMapping("/tables/schema/{schemaId}")
-    public Mono<BaseResponse<List<TableResponse>>> getTablesBySchemaId(
-            @PathVariable String schemaId) {
-        return tableService.getTablesBySchemaId(schemaId)
+    @GetMapping("/tables/{tableId}/columns")
+    public Mono<BaseResponse<List<ColumnResponse>>> getColumnsByTableId(
+            @PathVariable String tableId) {
+        return columnService.getColumnsByTableId(tableId)
+                .collectList()
+                .map(BaseResponse::success);
+    }
+
+    @PreAuthorize("hasAnyRole('OWNER','ADMIN','EDITOR','COMMENTER','VIEWER')")
+    @GetMapping("/tables/{tableId}/relationships")
+    public Mono<BaseResponse<List<RelationshipResponse>>> getRelationshipsByTableId(
+            @PathVariable String tableId) {
+        return relationshipService.getRelationshipsByTableId(tableId)
+                .collectList()
+                .map(BaseResponse::success);
+    }
+
+    @PreAuthorize("hasAnyRole('OWNER','ADMIN','EDITOR','COMMENTER','VIEWER')")
+    @GetMapping("/tables/{tableId}/indexes")
+    public Mono<BaseResponse<List<IndexResponse>>> getIndexesByTableId(
+            @PathVariable String tableId) {
+        return indexService.getIndexesByTableId(tableId)
+                .collectList()
+                .map(BaseResponse::success);
+    }
+
+    @PreAuthorize("hasAnyRole('OWNER','ADMIN','EDITOR','COMMENTER','VIEWER')")
+    @GetMapping("/tables/{tableId}/constraints")
+    public Mono<BaseResponse<List<ConstraintResponse>>> getConstraintsByTableId(
+            @PathVariable String tableId) {
+        return constraintService.getConstraintsByTableId(tableId)
                 .collectList()
                 .map(BaseResponse::success);
     }
