+++ apps/backend/core/src/test/java/com/schemafy/core/erd/controller/ConstraintControllerTest.java
@@ -1,7 +1,6 @@
 package com.schemafy.core.erd.controller;
 
 import java.util.Collections;
-import java.util.List;
 import java.util.Map;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,7 +16,6 @@ import org.springframework.test.web.reactive.server.WebTestClient;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.protobuf.Message;
 import com.google.protobuf.util.JsonFormat;
@@ -27,7 +25,6 @@ import com.schemafy.core.erd.controller.dto.response.AffectedMappingResponse;
 import com.schemafy.core.erd.controller.dto.response.ConstraintResponse;
 import com.schemafy.core.erd.service.ConstraintService;
 
-import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 import validation.Validation;
 
@@ -169,7 +166,7 @@ class ConstraintControllerTest {
 
         when(constraintService.createConstraint(
                 any(Validation.CreateConstraintRequest.class)))
-                .thenReturn(Mono.just(mockResponse));
+                        .thenReturn(Mono.just(mockResponse));
 
         webTestClient.post()
                 .uri(API_BASE_PATH + "/constraints")
@@ -234,7 +231,7 @@ class ConstraintControllerTest {
                                         .description("전파된 컬럼 목록"),
                                 fieldWithPath(
                                         "result.propagated.constraintColumns")
-                                        .description("전파된 제약조건 컬럼 목록"),
+                                                .description("전파된 제약조건 컬럼 목록"),
                                 fieldWithPath("result.propagated.indexColumns")
                                         .description("전파된 인덱스 컬럼 목록"))));
     }
@@ -303,78 +300,6 @@ class ConstraintControllerTest {
                                 fieldWithPath("result.columns[].columnId")
                                         .description("컬럼 ID"),
                                 fieldWithPath("result.columns[].seqNo")
-                                        .description("순서 번호"))));
-    }
-
-    @Test
-    @DisplayName("테이블별 제약조건 목록 조회 API 문서화")
-    void getConstraintsByTableId() throws Exception {
-        String mockResponseJson = """
-                [
-                    {
-                        "id": "06D6WWYRQCEXN1ACZ4ZJ12DFTG",
-                        "tableId": "06D6W8HDY79QFZX39RMX62KSX4",
-                        "name": "PK",
-                        "kind": "CONSTRAINT_KIND_UNSPECIFIED",
-                        "columns": [
-                            {
-                                "id": "06D6WWYRTER122SPQBVSX4T3MR",
-                                "constraintId": "06D6WWYRQCEXN1ACZ4ZJ12DFTG",
-                                "columnId": "06D6W90RSE1VPFRMM4XPKYGM9M",
-                                "seqNo": 1
-                            }
-                        ]
-                    }
-                ]
-                """;
-
-        when(constraintService
-                .getConstraintsByTableId("06D6W8HDY79QFZX39RMX62KSX4"))
-                .thenReturn(Flux
-                        .fromIterable(objectMapper.readValue(mockResponseJson,
-                                new TypeReference<List<ConstraintResponse>>() {
-                                })));
-
-        webTestClient.get()
-                .uri(API_BASE_PATH + "/constraints/table/{tableId}",
-                        "06D6W8HDY79QFZX39RMX62KSX4")
-                .header("Accept", "application/json")
-                .exchange()
-                .expectStatus().isOk()
-                .expectBody()
-                .jsonPath("$.success").isEqualTo(true)
-                .jsonPath("$.result").isArray()
-                .jsonPath("$.result.length()").isEqualTo(1)
-                .consumeWith(document("constraint-get-by-table",
-                        pathParameters(
-                                parameterWithName("tableId")
-                                        .description("테이블 ID")),
-                        requestHeaders(
-                                headerWithName("Accept").description("응답 포맷")),
-                        responseHeaders(
-                                headerWithName("Content-Type")
-                                        .description("응답 컨텐츠 타입")),
-                        responseFields(
-                                fieldWithPath("success")
-                                        .description("요청 성공 여부"),
-                                fieldWithPath("result").description("제약조건 목록"),
-                                fieldWithPath("result[].id")
-                                        .description("제약조건 ID"),
-                                fieldWithPath("result[].tableId")
-                                        .description("테이블 ID"),
-                                fieldWithPath("result[].name")
-                                        .description("제약조건 이름"),
-                                fieldWithPath("result[].kind")
-                                        .description("제약조건 종류"),
-                                fieldWithPath("result[].columns")
-                                        .description("제약조건 컬럼 목록"),
-                                fieldWithPath("result[].columns[].id")
-                                        .description("제약조건 컬럼 ID"),
-                                fieldWithPath("result[].columns[].constraintId")
-                                        .description("제약조건 ID"),
-                                fieldWithPath("result[].columns[].columnId")
-                                        .description("컬럼 ID"),
-                                fieldWithPath("result[].columns[].seqNo")
                                         .description("순서 번호"))));
     }
 
@@ -476,8 +401,9 @@ class ConstraintControllerTest {
 
         when(constraintService.updateConstraintName(
                 any(Validation.ChangeConstraintNameRequest.class)))
-                .thenReturn(Mono.just(objectMapper.readValue(mockResponseJson,
-                        ConstraintResponse.class)));
+                        .thenReturn(Mono
+                                .just(objectMapper.readValue(mockResponseJson,
+                                        ConstraintResponse.class)));
 
         webTestClient.put()
                 .uri(API_BASE_PATH + "/constraints/{constraintId}/name",
@@ -612,21 +538,22 @@ class ConstraintControllerTest {
 
         when(constraintService.addColumnToConstraint(
                 any(Validation.AddColumnToConstraintRequest.class)))
-                .thenReturn(Mono.just(new AffectedMappingResponse(
-                        Collections.emptyMap(),
-                        Collections.emptyMap(),
-                        Collections.emptyMap(),
-                        Collections.emptyMap(),
-                        Collections.emptyMap(),
-                        Map.of("06D6W8HDY79QFZX39RMX62KSX4",
-                                Map.of("06D5XSF8RRRKMCHVNX68TDX1K4",
-                                        "06D6WWYRQCEXN1ACZ4ZJ12DFTG")),
-                        Map.of("06D6WWYRQCEXN1ACZ4ZJ12DFTG",
-                                Map.of("06D6X1234567890ABCDEFGHIJK",
-                                        "06D6X2345678901BCDEFGHIJKL")),
-                        Collections.emptyMap(),
-                        Collections.emptyMap(),
-                        AffectedMappingResponse.PropagatedEntities.empty())));
+                        .thenReturn(Mono.just(new AffectedMappingResponse(
+                                Collections.emptyMap(),
+                                Collections.emptyMap(),
+                                Collections.emptyMap(),
+                                Collections.emptyMap(),
+                                Collections.emptyMap(),
+                                Map.of("06D6W8HDY79QFZX39RMX62KSX4",
+                                        Map.of("06D5XSF8RRRKMCHVNX68TDX1K4",
+                                                "06D6WWYRQCEXN1ACZ4ZJ12DFTG")),
+                                Map.of("06D6WWYRQCEXN1ACZ4ZJ12DFTG",
+                                        Map.of("06D6X1234567890ABCDEFGHIJK",
+                                                "06D6X2345678901BCDEFGHIJKL")),
+                                Collections.emptyMap(),
+                                Collections.emptyMap(),
+                                AffectedMappingResponse.PropagatedEntities
+                                        .empty())));
 
         webTestClient.post()
                 .uri(API_BASE_PATH + "/constraints/{constraintId}/columns",
@@ -693,7 +620,7 @@ class ConstraintControllerTest {
                                         .description("전파된 컬럼 목록"),
                                 fieldWithPath(
                                         "result.propagated.constraintColumns")
-                                        .description("전파된 제약조건 컬럼 목록"),
+                                                .description("전파된 제약조건 컬럼 목록"),
                                 fieldWithPath("result.propagated.indexColumns")
                                         .description("전파된 인덱스 컬럼 목록"))));
     }
@@ -792,7 +719,7 @@ class ConstraintControllerTest {
 
         when(constraintService.removeColumnFromConstraint(
                 any(Validation.RemoveColumnFromConstraintRequest.class)))
-                .thenReturn(Mono.empty());
+                        .thenReturn(Mono.empty());
 
         webTestClient.method(org.springframework.http.HttpMethod.DELETE)
                 .uri(API_BASE_PATH
@@ -900,7 +827,7 @@ class ConstraintControllerTest {
 
         when(constraintService.deleteConstraint(
                 any(Validation.DeleteConstraintRequest.class)))
-                .thenReturn(Mono.empty());
+                        .thenReturn(Mono.empty());
 
         webTestClient.method(org.springframework.http.HttpMethod.DELETE)
                 .uri(API_BASE_PATH + "/constraints/{constraintId}",
