+++ apps/backend/core/src/test/java/com/schemafy/core/erd/controller/SchemaControllerTest.java
@@ -20,10 +20,15 @@ import com.google.protobuf.util.JsonFormat;
 import com.schemafy.core.common.constant.ApiPath;
 import com.schemafy.core.common.security.WithMockCustomUser;
 import com.schemafy.core.erd.controller.dto.response.AffectedMappingResponse;
+import com.schemafy.core.erd.controller.dto.response.MemoResponse;
 import com.schemafy.core.erd.controller.dto.response.SchemaDetailResponse;
 import com.schemafy.core.erd.controller.dto.response.SchemaResponse;
+import com.schemafy.core.erd.controller.dto.response.TableResponse;
+import com.schemafy.core.erd.service.MemoService;
 import com.schemafy.core.erd.service.SchemaService;
+import com.schemafy.core.erd.service.TableService;
 
+import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 import validation.Validation;
 
@@ -59,6 +64,12 @@ class SchemaControllerTest {
 
     @MockitoBean
     private SchemaService schemaService;
+
+    @MockitoBean
+    private TableService tableService;
+
+    @MockitoBean
+    private MemoService memoService;
 
     private String toJson(Message message) throws Exception {
         return JsonFormat.printer()
@@ -121,7 +132,7 @@ class SchemaControllerTest {
 
         given(schemaService
                 .createSchema(any(Validation.CreateSchemaRequest.class)))
-                .willReturn(Mono.just(mockResponse));
+                        .willReturn(Mono.just(mockResponse));
 
         webTestClient.post()
                 .uri(API_BASE_PATH + "/schemas")
@@ -174,7 +185,7 @@ class SchemaControllerTest {
                                         .description("전파된 컬럼 목록"),
                                 fieldWithPath(
                                         "result.propagated.constraintColumns")
-                                        .description("전파된 제약조건 컬럼 목록"),
+                                                .description("전파된 제약조건 컬럼 목록"),
                                 fieldWithPath("result.propagated.indexColumns")
                                         .description("전파된 인덱스 컬럼 목록"))));
     }
@@ -483,6 +494,139 @@ class SchemaControllerTest {
                 .expectStatus().isBadRequest()
                 .expectBody()
                 .jsonPath("$.success").isEqualTo(false);
+    }
+
+    @Test
+    @DisplayName("스키마별 테이블 목록 조회 API 문서화")
+    void getTablesBySchemaId() throws Exception {
+        String schemaId = "06D6VZBWHSDJBBG0H7D156YZ98";
+
+        TableResponse tableResponse = objectMapper.treeToValue(
+                objectMapper
+                        .readTree(
+                                """
+                                        {
+                                            "id": "06D6W1GAHD51T5NJPK29Q6BCR8",
+                                            "schemaId": "06D6VZBWHSDJBBG0H7D156YZ98",
+                                            "name": "users",
+                                            "comment": "사용자 테이블",
+                                            "tableOptions": "ENGINE=InnoDB DEFAULT CHARSET=utf8mb4",
+                                            "extra": "{}",
+                                            "createdAt": "2025-11-10T13:48:01Z",
+                                            "updatedAt": "2025-11-10T13:48:01Z"
+                                        }
+                                        """),
+                TableResponse.class);
+
+        given(tableService.getTablesBySchemaId(schemaId))
+                .willReturn(Flux.just(tableResponse));
+
+        webTestClient.get()
+                .uri(API_BASE_PATH + "/schemas/{schemaId}/tables", schemaId)
+                .header("Accept", "application/json")
+                .exchange()
+                .expectStatus().isOk()
+                .expectBody()
+                .jsonPath("$.success").isEqualTo(true)
+                .jsonPath("$.result").isArray()
+                .jsonPath("$.result.length()").isEqualTo(1)
+                .jsonPath("$.result[0].id")
+                .isEqualTo("06D6W1GAHD51T5NJPK29Q6BCR8")
+                .jsonPath("$.result[0].schemaId")
+                .isEqualTo("06D6VZBWHSDJBBG0H7D156YZ98")
+                .jsonPath("$.result[0].name").isEqualTo("users")
+                .jsonPath("$.result[0].comment").isEqualTo("사용자 테이블")
+                .jsonPath("$.result[0].tableOptions")
+                .isEqualTo("ENGINE=InnoDB DEFAULT CHARSET=utf8mb4")
+                .consumeWith(document("table-list-by-schema",
+                        pathParameters(
+                                parameterWithName("schemaId")
+                                        .description("조회할 스키마의 ID")),
+                        requestHeaders(
+                                headerWithName("Accept")
+                                        .description(
+                                                "응답 포맷 (application/json)")),
+                        responseHeaders(
+                                headerWithName("Content-Type")
+                                        .description("응답 컨텐츠 타입")),
+                        responseFields(
+                                fieldWithPath("success")
+                                        .description("요청 성공 여부"),
+                                fieldWithPath("result")
+                                        .description("테이블 목록"),
+                                fieldWithPath("result[].id")
+                                        .description("테이블 ID"),
+                                fieldWithPath("result[].schemaId")
+                                        .description("스키마 ID"),
+                                fieldWithPath("result[].name")
+                                        .description("테이블 이름"),
+                                fieldWithPath("result[].comment")
+                                        .description("테이블 설명"),
+                                fieldWithPath("result[].tableOptions")
+                                        .description("테이블 옵션"),
+                                fieldWithPath("result[].extra")
+                                        .description("추가 정보"),
+                                fieldWithPath("result[].createdAt")
+                                        .description("생성 일시"),
+                                fieldWithPath("result[].updatedAt")
+                                        .description("수정 일시"))));
+    }
+
+    @Test
+    @DisplayName("스키마별 메모 목록 조회 API 문서화")
+    void getMemosBySchemaId() throws Exception {
+        String schemaId = "06D6VZBWHSDJBBG0H7D156YZ98";
+
+        MemoResponse response = objectMapper.treeToValue(
+                objectMapper.readTree("""
+                        {
+                            "id": "06D6W1GAHD51T5NJPK29Q6BCR8",
+                            "schemaId": "06D6VZBWHSDJBBG0H7D156YZ98",
+                            "authorId": "01ARZ3NDEKTSV4RRFFQ69G5FAV",
+                            "positions": "{}",
+                            "createdAt": "2025-11-23T10:00:00Z",
+                            "updatedAt": "2025-11-23T10:00:00Z"
+                        }
+                        """), MemoResponse.class);
+
+        given(memoService.getMemosBySchemaId(schemaId))
+                .willReturn(Flux.just(response));
+
+        webTestClient.get()
+                .uri(API_BASE_PATH + "/schemas/{schemaId}/memos", schemaId)
+                .header("Accept", "application/json")
+                .exchange()
+                .expectStatus().isOk()
+                .expectBody()
+                .jsonPath("$.success").isEqualTo(true)
+                .consumeWith(document("memo-list-by-schema",
+                        pathParameters(
+                                parameterWithName("schemaId")
+                                        .description("조회할 스키마 ID")),
+                        requestHeaders(
+                                headerWithName("Accept")
+                                        .description(
+                                                "응답 포맷 (application/json)")),
+                        responseHeaders(
+                                headerWithName("Content-Type")
+                                        .description("응답 컨텐츠 타입")),
+                        responseFields(
+                                fieldWithPath("success")
+                                        .description("요청 성공 여부"),
+                                fieldWithPath("result")
+                                        .description("메모 목록"),
+                                fieldWithPath("result[].id")
+                                        .description("메모 ID"),
+                                fieldWithPath("result[].schemaId")
+                                        .description("스키마 ID"),
+                                fieldWithPath("result[].authorId")
+                                        .description("작성자 ID"),
+                                fieldWithPath("result[].positions")
+                                        .description("메모 위치"),
+                                fieldWithPath("result[].createdAt")
+                                        .description("생성 일시"),
+                                fieldWithPath("result[].updatedAt")
+                                        .description("수정 일시"))));
     }
 
 }
